+++ Assets/Scripts/GameLogic/Monster.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using UnityEngine;
+
+namespace Ventura.GameLogic
+{
+    [Serializable]
+    public class Monster : Actor, ISerializationCallbackReceiver
+    {
+        public Monster(string name) : base(name)
+        {
+            //_ai = new RandomMovementAI(this);
+        }
+    }
+}
+
+
+