+++ CHANGELOG.md
@@ -7,11 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- item generation in MapGenerator
+- item generation and drawing in map
+- (static) actor generation and drawing in in map
 - azione PickupItem
 - zoom in/out in map
 - A* pathfinding implementation
 - theme previewer (first draft)
+- water terrain
 
 ### Changed
 - Extracted SystemManager from GameManager to remove dirty DontDestroyOnLoad hack